Move ServerGoalHandle to separate module
