+++ rclrs/src/action.rs
@@ -1,17 +1,12 @@
 mod client;
 mod server;
+mod server_goal_handle;
 
-use crate::{rcl_bindings::*, RclrsError};
-use std::{marker::PhantomData, sync::Arc};
+use crate::rcl_bindings::RCL_ACTION_UUID_SIZE;
 
 pub use client::{ActionClient, ActionClientBase};
 pub use server::{ActionServer, ActionServerBase};
-
-// SAFETY: The functions accessing this type, including drop(), shouldn't care about the thread
-// they are running in. Therefore, this type can be safely sent to another thread.
-unsafe impl Send for rcl_action_goal_handle_t {}
-
-unsafe impl Sync for rcl_action_goal_handle_t {}
+pub use server_goal_handle::ServerGoalHandle;
 
 pub type GoalUUID = [u8; RCL_ACTION_UUID_SIZE];
 
@@ -24,46 +19,4 @@ pub enum GoalResponse {
 pub enum CancelResponse {
     Reject = 1,
     Accept = 2,
-}
-
-pub struct ServerGoalHandle<T>
-where
-    T: rosidl_runtime_rs::Action,
-{
-    rcl_handle: Arc<rcl_action_goal_handle_t>,
-    goal_request: Arc<T>,
-    _marker: PhantomData<T>,
-}
-
-impl<T> ServerGoalHandle<T>
-where
-    T: rosidl_runtime_rs::Action,
-{
-    pub fn new(rcl_handle: Arc<rcl_action_goal_handle_t>, goal_request: Arc<T>) -> Self {
-        Self {
-            rcl_handle,
-            goal_request: Arc::clone(&goal_request),
-            _marker: Default::default(),
-        }
-    }
-
-    pub fn is_canceling(&self) -> bool {
-        false
-    }
-
-    pub fn is_active(&self) -> bool {
-        false
-    }
-
-    pub fn is_executing(&self) -> bool {
-        false
-    }
-
-    pub fn succeed(&self, result: &T::Result) -> Result<(), RclrsError> {
-        Ok(())
-    }
-
-    pub fn canceled(&self, result: &T::Result) -> Result<(), RclrsError> {
-        Ok(())
-    }
 }
